+++ backend-services/voting-app/src/main/java/org/cardano/foundation/voting/repository/VoteMerkleProofRepository.java
@@ -12,7 +12,7 @@ import java.util.Optional;
 @Repository
 public interface VoteMerkleProofRepository extends JpaRepository<VoteMerkleProof, String> {
 
-    @Query("SELECT vmp FROM VoteMerkleProof vmp WHERE vmp.eventId = :eventId AND vmp.voteId = :voteId ORDER BY vmp.absoluteSlot, vmp.createdAt DESC")
+    @Query("SELECT vmp FROM VoteMerkleProof vmp WHERE vmp.eventId = :eventId AND vmp.voteId = :voteId ORDER BY vmp.absoluteSlot, vmp.voteIdNumericHash DESC")
     Optional<VoteMerkleProof> findLatestProof(@Param("eventId") String eventId, @Param("voteId") String voteId);
 
     @Query("UPDATE VoteMerkleProof vmp SET vmp.invalidated = true where vmp.absoluteSlot > :slot")